handle user screen term not starting w/ 'screen'
When the TERM env var in the screen session doesn't start with 'screen'
screen.vim will continuously think it needs to start a screen session,
leading to a infinite loop that only ends when ther user's terminal runs
out of screen space for vim to start.  Resolve this issue by checking
g:ScreenShellBootstrapped and only start a new screen session if not
set.  This won't prevent the creation of a nested session in an existing
screen session, but will behave much better.