create per impl functions to determine if already in a session
when using tmux it is easier and more reliable to determining if the
user is already in an active session without hoping that they are using
a screen* TERM and that they haven't set that TERM outside of a session.
So instead of checking TERM for both screen and tmux, only use TERM for
gnu screen and instead use the TMUX env var for tmux.

fixes #25